[#1530] make internal API rpc code shared between apps
