+++ apps/beets-frontend-v3/app/api/tenderly/gas-estimate/route.ts
@@ -1,13 +1,15 @@
 import type { NextRequest } from 'next/server'
 import {
   getTenderlyConfig,
-  getAllowedOrigins,
-  isAllowedOrigin,
   validateRequestBody,
   getGasEstimateSimulation,
-  createForbiddenResponse,
+} from '@repo/lib/shared/utils/internal-api/tenderly-gas-estimate'
+import {
   createBadRequestResponse,
-} from '@repo/lib/shared/utils/tenderly-gas-estimate'
+  createForbiddenResponse,
+  getAllowedOrigins,
+  isAllowedOrigin,
+} from '@repo/lib/shared/utils/internal-api/errors'
 
 const allowedOrigins = getAllowedOrigins()
 const tenderlyConfig = getTenderlyConfig()
