+++ packages/lib/shared/utils/internal-api/tenderly-gas-estimate.ts
@@ -1,5 +1,3 @@
-import type { NextRequest } from 'next/server'
-
 export interface TenderlyConfig {
   accountSlug: string | undefined
   projectSlug: string | undefined
@@ -25,18 +23,6 @@ export function getTenderlyConfig(): TenderlyConfig {
     projectSlug: process.env.NEXT_PRIVATE_TENDERLY_PROJECT_SLUG,
     accessKey: process.env.NEXT_PRIVATE_TENDERLY_ACCESS_KEY,
   }
-}
-
-export function getAllowedOrigins(): string[] {
-  return [
-    ...(process.env.NEXT_PRIVATE_ALLOWED_ORIGINS || '').split(','),
-    process.env.VERCEL_BRANCH_URL ? `https://${process.env.VERCEL_BRANCH_URL}` : '',
-  ].filter(Boolean)
-}
-
-export function isAllowedOrigin(request: NextRequest, allowedOrigins: string[]): boolean {
-  const referer = request.headers.get('referer')
-  return !!(referer && allowedOrigins.some(origin => referer.startsWith(origin)))
 }
 
 export function validateTenderlyConfig(config: TenderlyConfig): boolean {
@@ -94,24 +80,4 @@ export async function getGasEstimateSimulation(
       error: 'Internal server error',
     }
   }
-}
-
-export function createForbiddenResponse(): Response {
-  return new Response(
-    JSON.stringify({
-      error: 'Forbidden: Access denied',
-      code: -32000,
-      message: 'Access denied',
-    }),
-    {
-      status: 403,
-      headers: {
-        'Content-Type': 'application/json',
-      },
-    }
-  )
-}
-
-export function createBadRequestResponse(message: string): Response {
-  return new Response(JSON.stringify({ error: message }), { status: 400 })
 }
